+++ src/Manifest.hpp
@@ -75,11 +75,11 @@ struct Profile {
   const std::vector<std::string> ldflags;
   const bool lto;
   const bool debug;
-  const std::size_t optLevel;
+  const std::uint8_t optLevel;
 
   Profile(
       std::vector<std::string> cxxflags, std::vector<std::string> ldflags,
-      bool lto, bool debug, std::size_t optLevel
+      bool lto, bool debug, std::uint8_t optLevel
   ) noexcept
       : cxxflags(std::move(cxxflags)), ldflags(std::move(ldflags)), lto(lto),
         debug(debug), optLevel(optLevel) {}
